Debugging

Every (mortal) modeler writes models that contain mistakes or are susceptible to numerical errors in their hunt for the perfect model. Debugging such errors is part of the modeling process, and ModelingToolkit includes some functionality that helps with this.

For example, consider an ODE model with "dangerous" functions (here √):

using ModelingToolkit, OrdinaryDiffEq
using ModelingToolkit: t_nounits as t, D_nounits as D

@variables u1(t) u2(t) u3(t)
eqs = [D(u1) ~ -√(u1), D(u2) ~ -√(u2), D(u3) ~ -√(u3)]
defaults = [u1 => 1.0, u2 => 2.0, u3 => 3.0]
@named sys = ODESystem(eqs, t; defaults)
sys = structural_simplify(sys)

\[ \begin{align} \frac{\mathrm{d} \mathtt{u1}\left( t \right)}{\mathrm{d}t} &= - \sqrt{\mathtt{u1}\left( t \right)} \\ \frac{\mathrm{d} \mathtt{u2}\left( t \right)}{\mathrm{d}t} &= - \sqrt{\mathtt{u2}\left( t \right)} \\ \frac{\mathrm{d} \mathtt{u3}\left( t \right)}{\mathrm{d}t} &= - \sqrt{\mathtt{u3}\left( t \right)} \end{align} \]

This problem causes the ODE solver to crash:

julia> prob = ODEProblem(sys, [], (0.0, 10.0), []);
julia> sol = solve(prob, Tsit5());┌ Warning: At t=2.001167437857666, dt was forced below floating point epsilon 4.440892098500626e-16, and step error estimate = NaN. Aborting. There is either an error in your model specification or the true solution is unstable (or the true solution can not be represented in the precision of Float64).
└ @ SciMLBase ~/.julia/packages/SciMLBase/tWwhl/src/integrator_interface.jl:623

This suggests that something went wrong, but not exactly what went wrong and where it did. In such situations, the debug_system function is helpful:

julia> dsys = debug_system(sys; functions = [sqrt]);
julia> dprob = ODEProblem(dsys, [], (0.0, 10.0), []);
julia> dsol = solve(dprob, Tsit5());ERROR: Function sqrt(u1(t)) errors with input
  u1(t) => -9.057596781074334e-5

Now we see that it crashed because u1 decreased so much that it became negative and outside the domain of the √ function. We could have figured that out ourselves, but it is not always so obvious for more complex models.

ModelingToolkit.debug_system — Function

debug_system(sys::AbstractSystem; functions = [log, sqrt, (^), /, inv, asin, acos], error_nonfinite = true)

Wrap functions in sys so any error thrown in them shows helpful symbolic-numeric information about its input. If error_nonfinite, functions that output nonfinite values (like Inf or NaN) also display errors, even though the raw function itself does not throw an exception (like 1/0). For example:

julia> sys = debug_system(complete(sys))

julia> prob = ODEProblem(sys, [0.0, 2.0], (0.0, 1.0))

julia> prob.f(prob.u0, prob.p, 0.0)
ERROR: Function /(1, sin(P(t))) output non-finite value Inf with input
  1 => 1
  sin(P(t)) => 0.0