Dependency Graphs

Types

Missing docstring.

Missing docstring for BipartiteGraph. Check Documenter's build log for details.

Utility functions for BiPartiteGraphs

Missing docstring.

Missing docstring for Base.isequal. Check Documenter's build log for details.

Functions for calculating dependency graphs

Missing docstring.

Missing docstring for equation_dependencies. Check Documenter's build log for details.

Missing docstring.

Missing docstring for asgraph. Check Documenter's build log for details.

Missing docstring.

Missing docstring for variable_dependencies. Check Documenter's build log for details.

Missing docstring.

Missing docstring for asdigraph. Check Documenter's build log for details.

Missing docstring.

Missing docstring for eqeq_dependencies. Check Documenter's build log for details.

Missing docstring.

Missing docstring for varvar_dependencies. Check Documenter's build log for details.

Miscellaneous

ModelingToolkit.map_variables_to_equations — Function

map_variables_to_equations(
    sys::ModelingToolkitBase.AbstractSystem;
    rename_dummy_derivatives
) -> Dict{Union{Num, SymbolicUtils.BasicSymbolicImpl.var"typeof(BasicSymbolicImpl)"{T} where T}, Equation}

Given a system that has been simplified via mtkcompile, return a Dict mapping variables of the system to equations that are used to solve for them. This includes observed variables.

Keyword Arguments

rename_dummy_derivatives: Whether to rename dummy derivative variable keys into their Differential forms. For example, this would turn the key yˍt(t) into Differential(t)(y(t)).