System construction

Internal API

The content on this page describes internal implementation details of ModelingToolkit. These APIs may change without notice.

A System is the fundamental data structure in MTK. It is a hierarchical data structure which contains all symbolic information about a particular numerical problem. A System in general does not constrain what sorts of information can and cannot be provided about the system. It merely enforces invariants required for the construction of a valid system. Different problem constructors validate that the system they receive is valid for generating code for that type of numerical problem.

Time-dependent and independent systems

The most broad categorization of a system is time-dependence. A system may represent dynamics that vary over time (or any independent variable in general) or a time-independent problem. Time-dependent systems include those for ODEs, SDEs, Jump processes, etc. Time-independent ones include nonlinear systems, optimization systems, and so on. Programmatically, this distinction is made via the iv::Union{SymbolicT, Nothing} field of the system. Union-splitting or type-asserting this field access is highly recommended. For example, if a function knows it operates on time-dependent systems, it should use ModelingToolkitBase.get_iv (sys)::SymbolicT when accessing the field to improve type-stability and the quality of the code generated by the Julia compiler. In general, these sorts of type-assertions on unstable fields are very useful for performant symbolic-numeric code.

Developer interfaces

The following APIs are versioned for ModelingToolkit developers and extension packages. They are not intended as the default end-user simplification workflow. Downstream code may call alias_elimination on System values, but must not add methods to it. SymScope is a closed metadata interface; use its documented concrete scope constructors without subtyping it.

ModelingToolkit.alias_elimination — Function

alias_elimination(sys)

Run the alias-elimination pass on sys.

This is a lower-level structural simplification pass used by ModelingToolkitBase.mtkcompile. End-user code should usually call mtkcompile instead.

Developer Interface

Downstream packages may call alias_elimination on System values, but must not add methods to it. Its TearingState implementation is not an extension point.

Arguments

sys: system to transform.

Returns

The alias-eliminated system.

Examples

using ModelingToolkit

@independent_variables t
@variables x(t)
@named sys = System([Differential(t)(x) ~ x / x], t)
equations(alias_elimination(sys))

ModelingToolkitBase.SymScope — Type

SymScope

Abstract metadata key for variable scoping and namespacing in a hierarchical system.

Developer Interface

SymScope is a closed ModelingToolkit interface. External packages must not subtype it; use the documented LocalScope, ParentScope, and GlobalScope constructors instead.

Examples

using ModelingToolkitBase

@variables x
LocalScope(x)

Contracts of the type

The inner constructor for System takes the fields in order. While the order of fields doesn't really matter, it is useful to group them along with similar fields in the struct. All fields should be given as concrete of a type as possible without making System parametric.

The system should never store wrapped types. Any symbolic expression anywhere in the system should always be of type SymbolicT. Storing Num, Arr{T, N}, etc. is invalid. In general, SymbolicT is a useful type-alias. Due to the Const variant of BasicSymbolic, it is often useful to store values of various different types and sizes as a SymbolicT.

All fields should have a docstring. Fields which are only intended for internal access should be annotated with $INTERNAL_FIELD_WARNING at the top of the docstring. Note that "internal" here doesn't mean only ModelingToolkitBase. Internal fields can also be used by ModelingToolkitTearing and ModelingToolkit.

Immutability and updating fields

Immutability of `System`

System should only ever be an immutable type with no type-parameters. Introducing type-parameters makes ensuring type-stability of most symbolic passes very difficult. In the future, if the struct grows large enough that passing it to functions is a significant slowdown (evidenced by benchmarks or Julia's lowered IR) it can be made a mutable struct with all const fields. This is imperative for ensuring correct behavior.

Almost all fields of System should be considered immutable. Internal mutation is also disallowed, with the specific exceptions of ModelingToolkitBase.get_initial_conditions(sys) and ModelingToolkitBase.get_guesses(sys). Outside of these two fields, no other field may be mutated. To change fields of a system, use Setfield.@set or Setfield.@set!. This may be changed in the future to Accessors.@set and Accessors.@reset respectively, without breaking backward compatibility. For example, to add a parameter p to a system, do

sys # The system
ps = copy(get_ps(sys)) # note the `copy`
push!(ps, p)
@set! sys.ps = ps

In case multiple such fields need to be set in quick succession, using ConstructionBase.setproperties directly generates better code after compilation. This function is implemented for AbstractSystem as an @generated function which performs additional validation/updates and is what @set/@set! eventually call.

Constructors

The inner constructor of System optionally performs some basic validation fundamental to the type, and builds the struct. The checks keyword is used to control what validation is performed. The checks here are not allowed to mutate the fields, and should not be unduly expensive to perform. This constructor is only ever called internally, and is not user-facing.

The main user-facing constructor is System(eqs::Vector{Equation}, iv, dvs, ps, brownians = SymbolicT[]; kws...). Most fields of System are provided as keyword arguments. This constructor performs significant additional validation.

!!! note __legacy_defaults__ This keyword arguments only exists for the now-deprecated @mtkmodel. It should not be used by anything else.

This constructor is responsible for:

Unwrapping all wrapped symbolic variables/expressions passed to it.

Coercing all containers to the appropriate type.

Populating var_to_name.

Validating bindings.

Parsing bindings/initial conditions/guesses from variable metadata.

Converting symbolic event shorthand into SymbolicContinuousCallback/SymbolicDiscreteCallback.

Parsing metadata

It is often useful to disable parsing bindings/initial conditions/guesses from variable metadata during system construction. This is usually when a new system is built from an existing one. Passing discover_from_metadata = false disables the parsing.

Other constructors are shorthand that call the main constructor. Notably, System(eqs, t) automatically discovers variables and parameters from the equations (and other information) passed to the constructor. Similar behavior exists for the time-independent constructors. The variable discovery process is very carefully crafted. Delays, BVP constraints, etc. all reference variables in non-standard ways which require precise handling.

System equality

System uses the default fallback of isequal, which is to compare referential equality. This is because checking if two expressions are equivalent is undecidable in general. In addition, it is very expensive to check equality in a way that is agnostic to variable/equation ordering. Base.isapprox is provided to check this sort of equality, in case it is necessary.

Base.isapprox — Method

isapprox(sysa::System, sysb::System) -> Bool

Check if two systems are about equal, to the extent that ModelingToolkitBase.jl supports. Note that if this returns true, the systems are not guaranteed to be exactly equivalent (unless sysa === sysb) but are highly likely to represent a similar mathematical problem. If this returns false, the systems are very likely to be different.

Metadata API

Systems allow storing and accessing metadata using the same API as symbolic expressions. SymbolicUtils.setmetadata, SymbolicUtils.getmetadata, and SymbolicUtils.hasmetadata are implemented appropriately. Note that setmetadata returns a new system with the added metadata, since Systems are immutable (similar to the behavior for expressions).

In addition, System provides a caching API. This cache is mutable, and storing values here does not return a new system. This cache is currently only for internal use. It is highly useful for caching expensive intermediate results, such as symbolic jacobians.

ModelingToolkitBase.check_mutable_cache — Function

check_mutable_cache(
    sys::System,
    K::DataType,
    _::Type{V},
    default
) -> Any

If the mutable cache of sys contains an entry for key K return it. Otherwise, return default. V is the expected type of the cache entry, if it exists.

ModelingToolkitBase.store_to_mutable_cache! — Function

store_to_mutable_cache!(sys::System, K::DataType, value)

Store into the mutable cache of sys the key-value pair K => value.

ModelingToolkitBase.should_invalidate_mutable_cache_entry — Function

should_invalidate_mutable_cache_entry(
    K::DataType,
    patch::NamedTuple
) -> Bool

Return a Bool indicating whether the cached entry associated with key K should be invalidated if the given patch is applied to the system. By default, entries will be invalidated on all patches.

These functions are only intended to be used on flattened systems. Behavior is undefined when used on non-flattened systems.

Preparation for codegen

Systems have additional constraints before codegen can be done. For differential-equation-based systems (ODEs, DAEs, SDEs, etc.) the equations must either be differential equations of the form

D(x) ~ rhs

or algebraic equations of the form

0 ~ rhs

Here, rhs is an expression that cannot involve operator applications such as D(x). rhs may involve observed variables, in which case the required observed equations will be present during code generation.

In addition, all systems must be flattened and marked as complete via ModelingToolkitBase.complete. The function will not validate the above structure of equations. It performs the following operations:

Discover all GlobalScope variables in the system.

Expand all connect equations.

Flatten the hierarchical system.

Convert discrete variables (@discretes) to parameters.

Construct ParameterBindingsGraph.

Remove bound parameters from the list of parameters

Validate the absence of any equations involving only parameters.

Add Initial parameters to the system.

Appropriately prepare all symbolic callbacks.

If split = true (recommended) creates the IndexCache which contains indexing and structure information for parameters. Also reorders all parameters as required by the data structure.

Disabling namespacing for the top-level system.

Note that this list may change between versions. In general, systems are allowed to perform any transformation and caching during ModelingToolkitBase.complete which does not modify the intent of the system.

Initial parameters

ModelingToolkit inserts special Initial parameters during ModelingToolkitBase.complete. These are an important utility for correctly building the initialization system during codegen. They are almost always invisible to the user. They do not show up in ModelingToolkitBase.parameters(sys), but can be obtained via ModelingToolkitBase.get_ps(sys). They are stored in the .initials field of MTKParameters.

Initial parameters exist for every unknown and observed. For time dependent systems, Initial parameters are also added for the first derivative of each unknown and observed variable. All floating-point discrete variables detected by ModelingToolkitBase.is_variable_floatingpoint also have an Initial parameter, as do floating-point bound parameters.

Observed equations must be topologically sorted. This is also required for ModelingToolkitBase.full_equations to work correctly.

SDEs must be phrased in terms of noise_eqs and not brownians. No brownian variables should be present in the equations.

The order and number of unknowns should correspond to that of equations. For example, if the ith equation is

D(x) ~ rhs

Then x should be the ith unknown. Algebraic variables often have a weaker correspondence.

AtomicArrayDict

ModelingToolkit stores variable-value mappings using AtomicArrayDict. It is a custom AbstractDict wrapping any other AbstractDict, and disallows scalarized/indexed arrays from being keys. Initial conditions, guesses and bindings are stored as this type (or wrappers thereof).

AtomicArrayDict, AtomicArraySet, and their helpers are private implementation details. They must not be used as extension points by downstream packages.

Variable utilities

get_all_discretes, get_all_discretes_fast, split_indexed_var, and get_stable_index support internal system transformations. They are not extension points. In particular, get_stable_index returns a SymbolicUtils.StableIndex{Int} for an indexed symbolic array variable.

Best practices

struct Schedule is defined alongside System for convenience, since it is stored in the system by ModelingToolkitTearing when running mtkcompile and contains useful information about the simplified system.

Always use @unpack to obtain variables/parameters/subsystems from a system you intend to extend.

Always use @named to call component constructors. All component constructors should use the @component or @connector macros as appropriate.

In case of conflicts, initial conditions and guesses from parent systems take priority over those in children.

Any container of symbolic variables/expressions should use SymbolicT as the corresponding type. Functions should be as concretely typed as possible, even to the extent of restricting inputs to exact types such as Vector{SymbolicT}. This helps improve type-stability, reduce invalidations, and improve precompilation. This is all the more true of internal functions. Such practice also helps identify bugs when unsupported types are passed into functions, or the impacts of changing a container type when it is passed downstream.