DiscreteSystem

System Constructors

ModelingToolkit.DiscreteSystem — Type

struct DiscreteSystem <: AbstractTimeDependentSystem

A system of difference equations.

Fields

tag: tag: a tag for the system. If two systems have the same tag, then they are structurally identical.

eqs: The differential equations defining the discrete system.

iv: Independent variable.

states: Dependent (state) variables. Must not contain the independent variable.

ps: Parameter variables. Must not contain the independent variable.

tspan: Time span.

var_to_name: Array variables.

ctrls: Control parameters (some subset of ps).

observed: Observed states.

name: Name: the name of the system

systems: systems: The internal systems. These are required to have unique names.

defaults: defaults: The default values to use when initial conditions and/or parameters are not supplied in DiscreteProblem.

preface: preface: inject assignment statements before the evaluation of the RHS function.

connector_type: connector_type: type of the system

metadata: metadata: metadata for the system, to be used by downstream packages.

gui_metadata: gui_metadata: metadata for MTK GUI.

tearing_state: tearing_state: cache for intermediate tearing state

substitutions: substitutions: substitutions generated by tearing.

complete: complete: if a model sys is complete, then sys.x no longer performs namespacing.

Example

using ModelingToolkit

@parameters σ=28.0 ρ=10.0 β=8/3 δt=0.1
@variables t x(t)=1.0 y(t)=0.0 z(t)=0.0
D = Difference(t; dt=δt)

eqs = [D(x) ~ σ*(y-x),
       D(y) ~ x*(ρ-z)-y,
       D(z) ~ x*y - β*z]

@named de = DiscreteSystem(eqs,t,[x,y,z],[σ,ρ,β]; tspan = (0, 1000.0)) # or
@named de = DiscreteSystem(eqs)

Composition and Accessor Functions

get_eqs(sys) or equations(sys): The equations that define the Discrete System.

get_delay_val(sys): The delay of the Discrete System.

get_iv(sys): The independent variable of the Discrete System.

get_u0_p(sys, u0map, parammap) Numeric arrays for the initial condition and parameters given var => value maps.

Transformations

Analyses

Applicable Calculation and Generation Functions

Standard Problem Constructors

SciMLBase.DiscreteFunction — Method

SciMLBase.DiscreteFunction{iip}(sys::DiscreteSystem, dvs = states(sys),
                                ps = parameters(sys);
                                version = nothing,
                                kwargs...) where {iip}

Create a DiscreteFunction from the DiscreteSystem. The arguments dvs and ps are used to set the order of the dependent variable and parameter vectors, respectively.

Missing docstring.

Missing docstring for DiscreteProblem(sys::ModelingToolkit.DiscreteSystem, args...). Check Documenter's build log for details.

Expression Constructors

ModelingToolkit.DiscreteFunctionExpr — Type

DiscreteFunctionExpr{iip}(sys::DiscreteSystem, dvs = states(sys),
                          ps = parameters(sys);
                          version = nothing,
                          kwargs...) where {iip}

Create a Julia expression for an DiscreteFunction from the DiscreteSystem. The arguments dvs and ps are used to set the order of the dependent variable and parameter vectors, respectively.

ModelingToolkit.DiscreteProblemExpr — Type

DiffEqBase.DiscreteProblemExpr(sys::JumpSystem, u0map, tspan,
                               parammap = DiffEqBase.NullParameters; kwargs...)

Generates a blank DiscreteProblem for a JumpSystem to utilize as its solving prob.prob. This is used in the case where there are no ODEs and no SDEs associated with the system.

Continuing the example from the JumpSystem definition:

using DiffEqBase, JumpProcesses
u₀map = [S => 999, I => 1, R => 0]
parammap = [β => 0.1 / 1000, γ => 0.01]
tspan = (0.0, 250.0)
dprob = DiscreteProblem(js, u₀map, tspan, parammap)