NonlinearSystem

System Constructors

ModelingToolkit.NonlinearSystem — Type

struct NonlinearSystem <: ModelingToolkit.AbstractSystem

A nonlinear system of equations.

Fields

eqs

Vector of equations defining the system.

states

Unknown variables.

ps

Parameters.

observed

name

Name: the name of the system

systems

systems: The internal systems

default_u0

default_u0: The default initial conditions to use when initial conditions are not supplied in ODEProblem.

default_p

default_p: The default parameters to use when parameters are not supplied in ODEProblem.

structure

structure: structural information of the system

reduced_states

Examples

@variables x y z
@parameters σ ρ β

eqs = [0 ~ σ*(y-x),
       0 ~ x*(ρ-z)-y,
       0 ~ x*y - β*z]
ns = NonlinearSystem(eqs, [x,y,z],[σ,ρ,β])

Composition and Accessor Functions

get_eqs(sys) or equations(sys): The equations that define the nonlinear system.

get_states(sys) or states(sys): The set of states in the nonlinear system.

get_ps(sys) or parameters(sys): The parameters of the nonlinear system.

Transformations

Missing docstring.

Missing docstring for structural_simplify. Check Documenter's build log for details.

Missing docstring.

Missing docstring for alias_elimination. Check Documenter's build log for details.

ModelingToolkit.StructuralTransformations.tearing — Function

tearing(sys; simplify=false)

Tear the nonlinear equations in system. When simplify=true, we simplify the new residual residual equations after tearing.

Analyses

Missing docstring.

Missing docstring for ModelingToolkit.islinear. Check Documenter's build log for details.

Applicable Calculation and Generation Functions

calculate_jacobian
generate_jacobian
jacobian_sparsity

Problem Constructors

SciMLBase.NonlinearProblem — Type

function DiffEqBase.NonlinearProblem{iip}(sys::NonlinearSystem,u0map,
                                          parammap=DiffEqBase.NullParameters();
                                          jac = false, sparse=false,
                                          checkbounds = false,
                                          linenumbers = true, parallel=SerialForm(),
                                          kwargs...) where iip

Generates an NonlinearProblem from a NonlinearSystem and allows for automatically symbolically calculating numerical enhancements.

Torn Problem Constructors

Missing docstring.

Missing docstring for BlockNonlinearProblem. Check Documenter's build log for details.